requirements.txt
aioredis==2.0.1
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
bcrypt==3.2.2
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cleo==1.0.0a4
click==8.0.4
colorama==0.4.4
crashtest==0.3.1
cryptography==37.0.2
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
Django==4.0.3
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
entrypoints==0.4
executing==0.8.3
fastapi==0.75.2
filelock==3.6.0
Flask==2.0.3
greenlet==1.1.2
h11==0.13.0
hiredis==2.0.0
httptools==0.4.0
idna==3.3
ipykernel==6.9.1
ipython==8.1.1
ipython-genutils==0.2.0
itsdangerous==2.1.1
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-server==1.13.5
jupyterlab-pygments==0.1.2
Mako==1.2.0
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.12
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
notebook-shim==0.1.0
orjson==3.6.8
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
platformdirs==2.5.1
pptree==3.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
Pygments==2.11.2
pylev==1.4.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
python-ulid==1.1.0
pytz==2021.3
pywin32==303
pywinpty==1.1.6
PyYAML==6.0
pyzmq==22.3.0
redis==4.2.2
redis-om==0.0.24
requests==2.27.1
rsa==4.8
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
sqlparse==0.4.2
stack-data==0.2.0
starlette==0.17.1
terminado==0.13.3
testpath==0.6.0
tornado==6.1
traitlets==5.1.1
types-redis==4.1.22
types-six==1.16.15
typing_extensions==4.2.0
tzdata==2022.1
ujson==5.3.0
urllib3==1.26.8
uvicorn==0.17.6
vboxapi==1.0
virtualenv==20.13.4
virtualenvwrapper-win==1.2.7
watchgod==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
websockets==10.3
Werkzeug==2.0.3
wrapt==1.14.0
